add create prompt examples, improve prompt manager

+++ src/genai/routers/prompt_saving.py
@@ -29,28 +29,32 @@ class PromptSavingRouter:
         except Exception as e:
             raise GenAiException(e)
 
-    def update_prompt(self, prompt_id: str, options: Options):
+    def update_prompt(self, id: str, options: Options):
         try:
-            endpoint = self.service_url + PromptSavingRouter.PROMPT_SAVING + "/" + prompt_id
+            endpoint = self.service_url + PromptSavingRouter.PROMPT_SAVING + "/" + id
             return RequestHandler.put(
                 endpoint,
                 key=self.key,
-                id=prompt_id,
+                id=id,
                 options=options,
             )
         except Exception as e:
             raise GenAiException(e)
 
-    def get_prompt(self, prompt_id: str):
+    def get_prompt(self, id: str):
+        if id is None:
+            raise GenAiException("ID cannot be None when attempting to get a prompt.")
         try:
-            endpoint = self.service_url + PromptSavingRouter.PROMPT_SAVING + "/" + prompt_id
+            endpoint = self.service_url + PromptSavingRouter.PROMPT_SAVING + "/" + id
             return RequestHandler.get(endpoint, key=self.key)
         except Exception as e:
             raise GenAiException(e)
 
-    def delete_prompt(self, prompt_id: str):
+    def delete_prompt(self, id: str):
+        if id is None:
+            raise GenAiException("ID cannot be None when attempting to delete a prompt.")
         try:
-            endpoint = self.service_url + PromptSavingRouter.PROMPT_SAVING + "/" + prompt_id
+            endpoint = self.service_url + PromptSavingRouter.PROMPT_SAVING + "/" + id
             return RequestHandler.delete(endpoint, key=self.key)
         except Exception as e:
             raise GenAiException(e)
